Compile paint code as cpp. Style fixes.
