+++ src/openrct2/paint/sprite/sprite.h
@@ -20,12 +20,20 @@
 #include "../../common.h"
 #include "../../world/sprite.h"
 
-typedef struct paint_session paint_session;
+#ifdef __cplusplus
+extern "C" {
+#endif 
 
-void sprite_paint_setup(paint_session * session, const uint16 eax, const uint16 ecx);
+    typedef struct paint_session paint_session;
 
-void misc_paint(paint_session * session, rct_sprite *misc, sint32 imageDirection);
-void litter_paint(paint_session * session, rct_litter *litter, sint32 imageDirection);
-void peep_paint(paint_session * session, rct_peep *peep, sint32 imageDirection);
+    void sprite_paint_setup(paint_session * session, const uint16 eax, const uint16 ecx);
+
+    void misc_paint(paint_session * session, rct_sprite *misc, sint32 imageDirection);
+    void litter_paint(paint_session * session, rct_litter *litter, sint32 imageDirection);
+    void peep_paint(paint_session * session, rct_peep *peep, sint32 imageDirection);
+
+#ifdef __cplusplus
+}
+#endif 
 
 #endif
